LPS-51313 Move the test
